Fix how trails work
The new trail length function measures the length of the trail by
percentage of a modified second. A modified second, as opposed to a
real-world second, is how long it takes for an arrow with a frequency of
1 to make a full rotation.

This new way of measuring trail length makes it more clear how the trail
will be, and no longer binds the length of the trail to the animation's
time scale.

I have also chosen to draw the trail at absolute increments of t = 0.001,
instead of at the points t - 0.001, t - 0.002, etc. This fixes some of
the vibrating edge effects that would occur, for example, in the
temperature rod seed.